+++ exec.c
@@ -23,7 +23,8 @@ exec(char *path, char **argv)
   pgdir = 0;
 
   // Check ELF header
-  if(readi(ip, (char*)&elf, 0, sizeof(elf)) < sizeof(elf))
+  // Ensure the full ELF header is read; cast to avoid signed/unsigned comparison.
+  if(readi(ip, (char*)&elf, 0, sizeof(elf)) < (int)sizeof(elf))
     goto bad;
   if(elf.magic != ELF_MAGIC)
     goto bad;
